完成http送来数据与前端的对接，能展示实时数据了

+++ src/main/java/cn/onebank/display/Top5Response.java
@@ -43,7 +43,7 @@ public class Top5Response implements Comparable<Top5Response> {
 
   @Override
   public int compareTo(Top5Response o) {
-    if (Long.parseLong(this.tps) >= Long.parseLong(o.tps)) {
+    if (Double.parseDouble(this.tps) >= Double.parseDouble(o.tps)) {
       return -1;
     } else {
       return 1;
